requirements.txt
streamlit
pandas
psycopg2-binary
openpyxl
plotly
python-dotenv
